requirements.txt
aiohttp
async-timeout
attrs
beautifulsoup4
cachetools
certifi
cffi
chardet
click
cryptography
discord.py
futures
google
google-api-core
google-api-python-client
google-assistant-grpc
google-assistant-sdk
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud
googleapis-common-protos
grpcio
httplib2
idna
multidict
oauthlib
pathlib2
pip
protobuf
pyOpenSSL
pyasn1
pyasn1-modules
pycparser
pytz
requests
requests-oauthlib
rsa
setuptools
six
sounddevice
soupsieve
tenacity
uritemplate
urllib3
websockets
yarl
python-firebase

